frontend/streamlit_app/requirements.txt
streamlit
langchain
langchain-openai
langchain-community
azure-search-documents==11.7.0b1
python-dotenv
azure-monitor-opentelemetry
langsmith
langfuse
